cloud-run/requirements.txt
functions-framework==3.*
pandas==2.1.3
google-cloud-bigquery==3.13.0
requests==2.31.0
pyarrow==14.0.1
fastapi==0.104.1
uvicorn==0.23.2